Make sure the header cell is always visible
When the collapsable data source opens, it will try to get all the cells showing on the screen. When the cells take up more room than the height of the table view, this used to push the header cell off. This implementation calculates the cell heights and determines which way it needs to scroll; This is probably not optimum and there maybe a better way. It should also be a method on the table view itself, to scroll to an array of index paths.