Fix issue where a stale cell was left on screen.
This fixes a situation where an extra cell is left in place on the table view. Hovering over the cell in the simulator results in "AX ERROR: Could not find my mock parent, most likely I am stale." printing to the log.

I believe this fixes it because it forces a reload of the header cell earlier, as it uses tableViewHasSetup to determine whether it can and this is not set to YES in numberOfSectionsInTableView:, the first call a UITableView makes to its dataSource.